Chocolate Bars
The Harris-Benedict equation estimates the number of calories your body needs to maintain your weight if you do no exercise. This is called your basal metabolic rate, or BMR.
The calories needed for a woman to maintain her weight is:
BMR = 655 + (4.3 x weight in pounds) + (4.7 x height in inches) – (4.7 x age in years)

The calories needed for a man to maintain his weight is:
BMR = 66 + (6.3 x weight in pounds) + (12.9 x height in inches) – (6.8 x age in years)

A typical chocolate bar will contain around 230 calories. Write a program that allows the user to input their weight in pounds, height in inches, and age in years. The program should then output the number of chocolate bars that should be consumed to maintain one’s weight for both a woman and a man of the input weight, height, and age.